[PGS] 나누어 떨어지는 숫자 배열 / Level 1 / 3분
- https://school.programmers.co.kr/learn/courses/30/lessons/12910
