+++ Seohyeon2/Seohyeon2/PGS/LEVEL_1/PGS_12910.swift
@@ -4,17 +4,23 @@
 //
 //  Created by seohyeon park on 2022/12/24.
 //
-// https://school.programmers.co.kr/learn/courses/30/lessons/12903
+// https://school.programmers.co.kr/learn/courses/30/lessons/12910
 
 import Foundation
 
-func answer_PGS_가운데_글자_가져오기_12903() {
-    func solution(_ s:String) -> String {
-        let arr = s.map { String($0) }
-        if arr.count % 2 == 0 {
-            return arr[(s.count / 2)-1] + arr[(s.count / 2)]
-        } else {
-            return arr[(s.count / 2)]
+func answer_PGS_나누어_떨어지는_숫자_배열_12910() {
+    func solution(_ arr:[Int], _ divisor:Int) -> [Int] {
+        var result = [Int]()
+        for number in arr {
+            if number % divisor == 0 {
+                result.append(number)
+            }
         }
+        
+        if !result.isEmpty {
+            return result.sorted()
+        }
+        
+        return [-1]
     }
 }
