+++ Seohyeon2/Seohyeon2/main.swift
@@ -136,6 +136,7 @@ answer_PGS_문자열_밀기_120921()
 /*
 answer_PGS_삼총사_131705()
 answer_PGS_2016년_12901()
+answer_PGS_가운데_글자_가져오기_12903()
 */
 
-answer_PGS_가운데_글자_가져오기_12903()
+answer_PGS_나누어_떨어지는_숫자_배열_12910()
